+++ client/scenes/project/scenes/survey/components/index.js
@@ -1,11 +1,2 @@
-export { default as CommentItem } from "./CommentItem/CommentItem";
-export { default as CommentModal } from "./CommentItem/CommentModal";
-export { default as OutstandingIssue } from "./CommentItem/OutstandingIssue";
-export { default as CommentBox } from "./CommentBox";
-export {
-  default as CommentBoxWithTagField
-} from "./CommentBox/CommentBoxWithTagField";
-export { default as EditCommentModal } from "./CommentItem/EditCommentModal";
 export { default as VersionToolbar } from "./VersionToolbar";
 export { default as SurveyHeader } from "./SurveyHeader";
-export { default as Countdown } from "./Countdown";
