+++ client/scenes/project/scenes/survey/scenes/Survey/components/CommentItem/ActionBar.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { PunditContainer, PunditTypeSet, VisibleIf } from "react-pundit";
-import policies from "../../../../../../policies.js";
+import policies from "../../../../../../../../policies.js";
 import { connect } from "react-redux";
 
 const ActionBar = ({
