+++ client/scenes/project/scenes/survey/scenes/Survey/components/CommentItem/CommentItem.scss
@@ -1,4 +1,4 @@
-@import "../../../../../../breakpoint";
+@import "../../../../../../../../breakpoint";
 
 .comment-item {
   background: #113954;
