+++ client/scenes/project/scenes/survey/scenes/Survey/components/index.js
@@ -7,3 +7,11 @@ export { default as SurveyIssues } from "./SurveyIssues";
 export { default as SidebarComments } from "./SidebarComments";
 export { default as SidebarTableOfContents } from "./SidebarTableOfContents";
 export { default as SurveyScorecard } from "./SurveyScorecard";
+export { default as CommentItem } from "./CommentItem/CommentItem";
+export { default as CommentModal } from "./CommentItem/CommentModal";
+export { default as CommentBox } from "./CommentBox";
+export {
+  default as CommentBoxWithTagField
+} from "./CommentBox/CommentBoxWithTagField";
+export { default as EditCommentModal } from "./CommentItem/EditCommentModal";
+export { default as Countdown } from "./Countdown";
