+++ client/scenes/project/scenes/survey/scenes/SurveyUpload/components/OutstandingIssue.scss
@@ -1,4 +1,4 @@
-@import "../../../../../../breakpoint";
+@import "../../../../../../../breakpoint";
 
 .comment-issue-item {
   background: transparent;
