+++ client/scenes/project/scenes/survey/scenes/SurveyUpload/components/index.js
@@ -1,2 +1,3 @@
 export { default as IssueInput } from "./IssueInput";
 export { default as UploadInterface } from "./UploadInterface";
+export { default as OutstandingIssue } from "./OutstandingIssue";
